back4/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
passlib
python-dotenv
pyjwt
